Only save to local storage when creating a new alert rule

+++ public/app/features/alerting/unified/components/rule-editor/alert-rule-form/AlertRuleForm.tsx
@@ -109,11 +109,13 @@ export const AlertRuleForm = ({ existing, prefill }: Props) => {
       notifyApp.error(conditionErrorMsg);
       return;
     }
-    // if this rule is using manual routing, we save this option in local storage
-    if (values.manualRouting) {
-      localStorage.setItem(MANUAL_ROUTING_KEY, 'true');
-    } else {
-      localStorage.setItem(MANUAL_ROUTING_KEY, 'false');
+    // when creating a new rule, we save the manual routing setting in local storage
+    if (!existing) {
+      if (values.manualRouting) {
+        localStorage.setItem(MANUAL_ROUTING_KEY, 'true');
+      } else {
+        localStorage.setItem(MANUAL_ROUTING_KEY, 'false');
+      }
     }
 
     dispatch(
